requirements.txt
Flask
Flask-CORS
rembg
Pillow
numpy
gunicorn
python-dotenv
Werkzeug
onnxruntime
